Add flag
