+++ server/src/http/mod.rs
@@ -1,6 +1,6 @@
 #[cfg(not(tarpaulin_include))]
 pub mod background_send_metrics;
 pub mod broadcaster;
-pub mod feature_refresher;
 pub(crate) mod headers;
 pub mod unleash_client;
+pub mod refresher;
